feat(dragdrop)!: remove top level subcomponents (#1697)
