--- src/components/DragDrop/DragDropTypes.ts
@@ -1,26 +1,0 @@
-import type React from 'react';
-
-export interface Items {
-  [key: string]: ItemType;
-}
-
-export interface ItemType {
-  id?: string;
-  title?: string;
-  behavior?: 'hover';
-  children?: React.ReactNode;
-  handle?: React.ReactNode;
-}
-
-export interface Containers {
-  [key: string]: ContainerType;
-}
-
-export interface ContainerType {
-  className?: string;
-  columnClassName?: string;
-  emptyContent?: React.ReactNode;
-  header?: React.ReactNode;
-  id?: string;
-  itemIds: string[];
-}
