--- src/components/DragDropContainer/index.ts
@@ -1,1 +1,0 @@
-export { DragDropContainer as default } from './DragDropContainer';
