--- src/components/DragDropContainerHeader/index.ts
@@ -1,1 +1,0 @@
-export { DragDropContainerHeader as default } from './DragDropContainerHeader';
