--- src/components/DragDropItem/index.ts
@@ -1,1 +1,0 @@
-export { DragDropItem as default } from './DragDropItem';
