+++ src/index.ts
@@ -14,9 +14,6 @@ export { default as Checkbox } from './components/Checkbox';
 export { default as ClickableStyle } from './components/ClickableStyle';
 export { default as DataBar } from './components/DataBar';
 export { default as DragDrop } from './components/DragDrop';
-export { default as DragDropContainer } from './components/DragDropContainer';
-export { default as DragDropContainerHeader } from './components/DragDropContainerHeader';
-export { default as DragDropItem } from './components/DragDropItem';
 export { default as Drawer } from './components/Drawer';
 export { default as Dropdown } from './components/Dropdown';
 export { default as DropdownButton } from './components/DropdownButton';
